+++ testSuite/imageFiltering/source/code/src/imageFilteringCUDA.cu
@@ -279,7 +279,7 @@ pmStatus imageFilter_cudaLaunchFunc(pmTaskInfo pTaskInfo, pmDeviceInfo pDeviceIn
 }
 
 // Returns 0 on success, non-zero on failure
-int singleGpuImageFilter(void* pInvertedImageData, int pImageWidth, int pImageHeight, char pFilter[MAX_FILTER_DIM][MAX_FILTER_DIM], int pFilterRadius, int pImageBytesPerLine, void* pOutputMem)
+int singleGpuImageFilter(void* pInvertedImageData, size_t pImageWidth, size_t pImageHeight, char pFilter[MAX_FILTER_DIM][MAX_FILTER_DIM], size_t pFilterRadius, size_t pImageBytesPerLine, void* pOutputMem)
 {
     void* lOutputMemCudaPtr = NULL;
     size_t lImageSize = (pImageWidth * pImageHeight * PIXEL_COUNT);
